TY: relax trait matching rules
